+++ app/src/main/java/com/lyfe/android/core/data/di/DataModule.kt
@@ -2,11 +2,13 @@ package com.lyfe.android.core.data.di
 
 import com.lyfe.android.core.data.repository.AlbumRepositoryImpl
 import com.lyfe.android.core.data.repository.AuthRepositoryImpl
+import com.lyfe.android.core.data.repository.TokenRepositoryImpl
 import com.lyfe.android.core.data.repository.UserRepositoryImpl
 import com.lyfe.android.core.data.repository.fake.FakeLyfeRepository
 import com.lyfe.android.core.domain.repository.AlbumRepository
 import com.lyfe.android.core.domain.repository.AuthRepository
 import com.lyfe.android.core.domain.repository.LyfeRepository
+import com.lyfe.android.core.domain.repository.TokenRepository
 import com.lyfe.android.core.domain.repository.UserRepository
 import dagger.Binds
 import dagger.Module
@@ -42,4 +44,10 @@ interface DataModule {
 	fun bindsAuthRepository(
 		authRepositoryImpl: AuthRepositoryImpl
 	): AuthRepository
+
+	@Singleton
+	@Binds
+	fun bindsTokenRepository(
+		tokenRepositoryImpl: TokenRepositoryImpl
+	): TokenRepository
 }
