+++ app/src/main/java/com/lyfe/android/core/domain/usecase/DeleteSignUpTokenUseCase.kt
@@ -1,11 +1,11 @@
 package com.lyfe.android.core.domain.usecase
 
-import com.lyfe.android.core.domain.repository.UserRepository
+import com.lyfe.android.core.domain.repository.TokenRepository
 import javax.inject.Inject
 
 class DeleteSignUpTokenUseCase @Inject constructor(
-	private val userRepository: UserRepository
+	private val tokenRepository: TokenRepository
 ) {
 
-	suspend operator fun invoke() = userRepository.deleteSignUpToken()
+	suspend operator fun invoke() = tokenRepository.deleteSignUpToken()
 }
