+++ app/src/main/java/com/lyfe/android/core/domain/usecase/GetRefreshTokenUseCase.kt
@@ -1,11 +1,11 @@
 package com.lyfe.android.core.domain.usecase
 
-import com.lyfe.android.core.domain.repository.UserRepository
+import com.lyfe.android.core.domain.repository.TokenRepository
 import javax.inject.Inject
 
 class GetRefreshTokenUseCase @Inject constructor(
-	private val userRepository: UserRepository
+	private val tokenRepository: TokenRepository
 ) {
 
-	operator fun invoke() = userRepository.getRefreshToken()
+	operator fun invoke() = tokenRepository.getRefreshToken()
 }
